Cast time_t may be signed on some systems so cast it so we don't get warning.

+++ licq/src/sighandler.c
@@ -131,7 +131,7 @@ void licq_handle_sigabrt(int s)
       FILE* file = fopen(filename, "a");
       if (file != NULL)
       {
-        fprintf(file, "\ntime: %lu\n", time(NULL));
+        fprintf(file, "\ntime: %lu\n", (unsigned long)time(NULL));
         fclose(file);
       }
